requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.2.0
streamlit>=1.10.0
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-learn>=1.0.0
tqdm>=4.62.3